+++ apps/ble/advanced_applications/ble_throughput/firmware/src/config/default/ble/service_ble/ble_dis/ble_dis.c
@@ -110,97 +110,97 @@ static const uint8_t s_chUuidPnpId[ATT_UUID_LENGTH_2] =                {UINT16_T
 #endif
 
 /* Device Information Service Declaration */
-static const uint16_t s_svcUuidDisLen = sizeof(s_svcUuidDis);
+static const uint16_t s_svcUuidDisLen = (uint16_t)sizeof(s_svcUuidDis);
 
 #ifdef DIS_MANU_NAME_ENABLE
 /* Manufacture Name Characteristic */
 static const uint8_t s_charManuName[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_MANU_NAME), UINT16_TO_BYTES(UUID_MANUFACTURER_NAME)};
-static const uint16_t s_charManuNameLen = sizeof(s_charManuName);
+static const uint16_t s_charManuNameLen = (uint16_t)sizeof(s_charManuName);
 
 /* Manufacture Name Value */
 static uint8_t s_manuNameVal[]={DIS_MANU_NAME};
-static uint16_t s_manuNameValLen = sizeof(s_manuNameVal);
+static uint16_t s_manuNameValLen = (uint16_t)sizeof(s_manuNameVal);
 #endif
 
 #ifdef DIS_MODEL_NUM_ENABLE
 /* Model Number Characteristic */
 static const uint8_t s_charModelNum[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_MODEL_NUM), UINT16_TO_BYTES(UUID_MODEL_NUMBER)};
-static const uint16_t s_charModelNumLen = sizeof(s_charModelNum);
+static const uint16_t s_charModelNumLen = (uint16_t)sizeof(s_charModelNum);
 
 /* Model Number Value */
 static uint8_t s_modelNumVal[] = {DIS_MODEL_NUM};
-static uint16_t s_modelNumValLen = sizeof(s_modelNumVal);
+static uint16_t s_modelNumValLen = (uint16_t)sizeof(s_modelNumVal);
 #endif
 
 #ifdef DIS_SERIAL_NUM_ENABLE
 /* Serial Number Characteristic */
 static const uint8_t s_charSerialNum[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_SERIAL_NUM), UINT16_TO_BYTES(UUID_SERIAL_NUMBER)};
-static const uint16_t s_charSerialNumLen = sizeof(s_charSerialNum);
+static const uint16_t s_charSerialNumLen = (uint16_t)sizeof(s_charSerialNum);
 
 /* Serial Number Value */
 static uint8_t s_serialNumVal[] = {DIS_SERIAL_NUM};
-static uint16_t s_serialNumValLen = sizeof(s_serialNumVal);
+static uint16_t s_serialNumValLen = (uint16_t)sizeof(s_serialNumVal);
 #endif
 
 #ifdef DIS_HW_REV_ENABLE
 /* Hardware Revision Characteristic */
 static const uint8_t s_charHwRev[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_HW_REV), UINT16_TO_BYTES(UUID_HARDWARE_REVISION)};
-static const uint16_t s_charHwRevLen = sizeof(s_charHwRev);
+static const uint16_t s_charHwRevLen = (uint16_t)sizeof(s_charHwRev);
 
 /* Hardware Revision Value */
 static uint8_t s_hwRevVal[] = {DIS_HW_REVISION};
-static uint16_t s_hwRevValLen = sizeof(s_hwRevVal);
+static uint16_t s_hwRevValLen = (uint16_t)sizeof(s_hwRevVal);
 #endif
 
 #ifdef DIS_FW_REV_ENABLE
 /* Firmware Revision Characteristic */
 static const uint8_t s_charFwRev[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_FW_REV), UINT16_TO_BYTES(UUID_FIRMWARE_REVISION)};
-static const uint16_t s_charFwRevLen = sizeof(s_charFwRev);
+static const uint16_t s_charFwRevLen = (uint16_t)sizeof(s_charFwRev);
 
 /* Firmware Revision Value */
 static uint8_t s_fwRevVal[] = {DIS_FW_REVISION};
-static uint16_t s_fwRevValLen = sizeof(s_fwRevVal);
+static uint16_t s_fwRevValLen = (uint16_t)sizeof(s_fwRevVal);
 #endif
 
 #ifdef DIS_SW_REV_ENABLE
 /* Software Revision Characteristic */
 static const uint8_t s_charSwRev[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_SW_REV), UINT16_TO_BYTES(UUID_SOFTWARE_REVISION)};
-static const uint16_t s_charSwRevLen = sizeof(s_charSwRev);
+static const uint16_t s_charSwRevLen = (uint16_t)sizeof(s_charSwRev);
 
 /* Software Revision Value */
 static uint8_t s_swRevVal[] = {DIS_SW_REVISION};
-static uint16_t s_swRevValLen = sizeof(s_swRevVal);
+static uint16_t s_swRevValLen = (uint16_t)sizeof(s_swRevVal);
 #endif
 
 #ifdef DIS_SYSTEM_ID_ENABLE
 /* System Id Characteristic */
 static const uint8_t s_charSystemId[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_SYSTEM_ID), UINT16_TO_BYTES(UUID_SYSTEM_ID)};
-static const uint16_t s_charSystemIdLen = sizeof(s_charSystemId);
+static const uint16_t s_charSystemIdLen = (uint16_t)sizeof(s_charSystemId);
 
 /* System Id Value */
 static uint8_t s_systemIdVal[] = {DIS_SYSTEM_ID};
-static uint16_t s_systemIdValLen = sizeof(s_systemIdVal);
+static uint16_t s_systemIdValLen = (uint16_t)sizeof(s_systemIdVal);
 #endif
 
 #ifdef DIS_IEEE_ENABLE
 /* IEEE 11073-20601 Regulatory Certification Data List Characteristic */
 static const uint8_t s_charIeee[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_IEEE_RCDL), UINT16_TO_BYTES(UUID_IEEE_11073_20601_RCDL)};
-static const uint16_t s_charIeeeLen = sizeof(s_charIeee);
+static const uint16_t s_charIeeeLen = (uint16_t)sizeof(s_charIeee);
 
 /* IEEE 11073-20601 Regulatory Certification Data List Value */
 static uint8_t s_ieeeVal[] = {DIS_IEEE_11073_20601};
-static uint16_t s_ieeeValLen = sizeof(s_ieeeVal);
+static uint16_t s_ieeeValLen = (uint16_t)sizeof(s_ieeeVal);
 #endif
 
 
 #ifdef DIS_PNP_ID_ENABLE
 /* PnpID Characteristic */
 static const uint8_t s_charPnpId[] = {(ATT_PROP_READ), UINT16_TO_BYTES(DIS_HDL_CHARVAL_PNP_ID), UINT16_TO_BYTES(UUID_PNP_ID)};
-static const uint16_t s_charPnpIdLen = sizeof(s_charPnpId);
+static const uint16_t s_charPnpIdLen = (uint16_t)sizeof(s_charPnpId);
 
 /* PnpID Value */
 static uint8_t s_pnpIdVal[] = {DIS_PNP_ID};
-static uint16_t s_pnpIdValLen = sizeof(s_pnpIdVal);
+static uint16_t s_pnpIdValLen = (uint16_t)sizeof(s_pnpIdVal);
 #endif
 
 /* Attribute list for Generic Access service */
@@ -226,7 +226,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidPrimSvc;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_svcUuidDis;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_svcUuidDisLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_svcUuidDis);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_svcUuidDis);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -236,7 +236,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charManuName;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charManuNameLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charManuName);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charManuName);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -244,7 +244,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidManuName;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_manuNameVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_manuNameValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_manuNameVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_manuNameVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -255,7 +255,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charModelNum;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charModelNumLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charModelNum);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charModelNum);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -263,7 +263,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidModelNum;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_modelNumVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_modelNumValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_modelNumVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_modelNumVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -274,7 +274,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charSerialNum;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charSerialNumLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charSerialNum);
+    s_disList[s_disAttrIndex].maxLen =(uint16_t) sizeof(s_charSerialNum);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -282,7 +282,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidSerialNum;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_serialNumVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_serialNumValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_serialNumVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_serialNumVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -293,7 +293,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charHwRev;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charHwRevLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charHwRev);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charHwRev);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -301,7 +301,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidHwRev;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_hwRevVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_hwRevValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_hwRevVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_hwRevVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -312,7 +312,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charFwRev;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charFwRevLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charFwRev);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charFwRev);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -320,7 +320,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidFwRev;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_fwRevVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_fwRevValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_fwRevVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_fwRevVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -331,7 +331,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charSwRev;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charSwRevLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charSwRev);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charSwRev);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -339,7 +339,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidSwRev;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_swRevVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_swRevValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_swRevVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_swRevVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -350,7 +350,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charSystemId;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charSystemIdLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charSystemId);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charSystemId);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -358,7 +358,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidSystemId;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_systemIdVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_systemIdValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_systemIdVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_systemIdVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -369,7 +369,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charIeee;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charIeeeLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charIeee);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charIeee);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -377,7 +377,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidIeeeRcdl;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_ieeeVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_ieeeValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_ieeeVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_ieeeVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -388,7 +388,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) g_gattUuidChar;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_charPnpId;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_charPnpIdLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_charPnpId);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_charPnpId);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
@@ -396,7 +396,7 @@ uint16_t BLE_DIS_Add(void)
     s_disList[s_disAttrIndex].p_uuid = (uint8_t *) s_chUuidPnpId;
     s_disList[s_disAttrIndex].p_value = (uint8_t *) s_pnpIdVal;
     s_disList[s_disAttrIndex].p_len = (uint16_t *) &s_pnpIdValLen;
-    s_disList[s_disAttrIndex].maxLen = sizeof(s_pnpIdVal);
+    s_disList[s_disAttrIndex].maxLen = (uint16_t)sizeof(s_pnpIdVal);
     s_disList[s_disAttrIndex].settings = 0;
     s_disList[s_disAttrIndex].permissions = PERMISSION_READ;
     s_disAttrIndex++;
